refactor: extract Android touch executor
